pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-background = "setuptools.build_meta"

[project]
name = "shopify-migration"
version = "1.0.0"
description = "Migrate products from CSV to Shopify format"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "cerberus>=1.3.4",
    "tqdm>=4.66.0",
    "colorama>=0.4.6",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

